Add Ability to Delete Words (#587)
* Add delete frontier word functionality
  - Frontier words are now deleted from the database and a new word is created in the word database
* Add accessibility to Word model
* Remove confusing delete icon background
* Make DeleteFrontierWord backend call return the new deleted word id
* Change state enum representation to string
* Mock cell components to fix failing tests
* Update delete functionality in DataEntry's delete function
* Use small size on sense deletion icon button for more uniform row heights
* Switch Delete icon to RestoreFromTrash when sense has been deleted.

+++ Backend.Tests/LiftControllerTests.cs
@@ -117,7 +117,7 @@ namespace Backend.Tests
 
             foreach (var sense in word.Senses)
             {
-                sense.Accessibility = (int)State.Active;
+                sense.Accessibility = State.Active;
                 sense.Glosses = new List<Gloss>() { new Gloss(), new Gloss(), new Gloss() };
 
                 foreach (var gloss in sense.Glosses)
